chore(hir,mir): clear what the recent reworks left behind
Needless borrows and imports the extern rework orphaned, an array
bound cast the i64 widening made redundant, and an enum test comment
that claimed a SINT base is one byte: it is a 4-byte slot like every
type of 32 bits or less, and the base changes the lane, not the slot.

+++ crates/mir/src/lower/lower_type.rs
@@ -274,7 +274,7 @@ fn lower_array_type<'db>(
     for (start_expr, end_expr) in array_type.subranges(db) {
         let start = extract_integer_literal(db, start_expr)?;
         let end = extract_integer_literal(db, end_expr)?;
-        dimensions.push((start as i64, end as i64));
+        dimensions.push((start, end));
 
         let dim_size = (end - start + 1).max(0) as u32;
         total_elements = total_elements.checked_mul(dim_size).ok_or_else(|| {
